Remove InstancesDao usage from InstanceUploaderTask
