--- collect_app/src/main/java/org/odk/collect/android/dao/InstancesDao.java
@@ -1,84 +1,0 @@
-/*
- * Copyright 2017 Nafundi
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.odk.collect.android.dao;
-
-import android.database.Cursor;
-
-import org.odk.collect.android.application.Collect;
-import org.odk.collect.android.instances.Instance;
-import org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns;
-
-import java.util.ArrayList;
-import java.util.List;
-
-/**
- * This class is used to encapsulate all access to the {@link org.odk.collect.android.provider.InstanceProvider#DATABASE_NAME}
- * For more information about this pattern go to https://en.wikipedia.org/wiki/Data_access_object
- */
-public class InstancesDao {
-
-    public Cursor getInstancesCursor(String[] projection, String selection, String[] selectionArgs, String sortOrder) {
-        return Collect.getInstance().getContentResolver().query(InstanceColumns.CONTENT_URI, projection, selection, selectionArgs, sortOrder);
-    }
-
-    /**
-     * Returns all instances available through the cursor and closes the cursor.
-     */
-    public static List<Instance> getInstancesFromCursor(Cursor cursor) {
-        List<Instance> instances = new ArrayList<>();
-        if (cursor != null) {
-            try {
-                cursor.moveToPosition(-1);
-                while (cursor.moveToNext()) {
-                    int displayNameColumnIndex = cursor.getColumnIndex(InstanceColumns.DISPLAY_NAME);
-                    int submissionUriColumnIndex = cursor.getColumnIndex(InstanceColumns.SUBMISSION_URI);
-                    int canEditWhenCompleteIndex = cursor.getColumnIndex(InstanceColumns.CAN_EDIT_WHEN_COMPLETE);
-                    int instanceFilePathIndex = cursor.getColumnIndex(InstanceColumns.INSTANCE_FILE_PATH);
-                    int jrFormIdColumnIndex = cursor.getColumnIndex(InstanceColumns.JR_FORM_ID);
-                    int jrVersionColumnIndex = cursor.getColumnIndex(InstanceColumns.JR_VERSION);
-                    int statusColumnIndex = cursor.getColumnIndex(InstanceColumns.STATUS);
-                    int lastStatusChangeDateColumnIndex = cursor.getColumnIndex(InstanceColumns.LAST_STATUS_CHANGE_DATE);
-                    int deletedDateColumnIndex = cursor.getColumnIndex(InstanceColumns.DELETED_DATE);
-                    int geometryTypeColumnIndex = cursor.getColumnIndex(InstanceColumns.GEOMETRY_TYPE);
-                    int geometryColumnIndex = cursor.getColumnIndex(InstanceColumns.GEOMETRY);
-
-                    int databaseIdIndex = cursor.getColumnIndex(InstanceColumns._ID);
-
-                    Instance instance = new Instance.Builder()
-                            .displayName(cursor.getString(displayNameColumnIndex))
-                            .submissionUri(cursor.getString(submissionUriColumnIndex))
-                            .canEditWhenComplete(Boolean.valueOf(cursor.getString(canEditWhenCompleteIndex)))
-                            .instanceFilePath(cursor.getString(instanceFilePathIndex))
-                            .jrFormId(cursor.getString(jrFormIdColumnIndex))
-                            .jrVersion(cursor.getString(jrVersionColumnIndex))
-                            .status(cursor.getString(statusColumnIndex))
-                            .lastStatusChangeDate(cursor.getLong(lastStatusChangeDateColumnIndex))
-                            .deletedDate(cursor.isNull(deletedDateColumnIndex) ? null : cursor.getLong(deletedDateColumnIndex))
-                            .geometryType(cursor.getString(geometryTypeColumnIndex))
-                            .geometry(cursor.getString(geometryColumnIndex))
-                            .id(cursor.getLong(databaseIdIndex))
-                            .build();
-
-                    instances.add(instance);
-                }
-            } finally {
-                cursor.close();
-            }
-        }
-        return instances;
-    }
-}
