+++ collect_app/src/main/java/org/odk/collect/android/tasks/InstanceUploaderTask.java
@@ -16,29 +16,22 @@
 
 package org.odk.collect.android.tasks;
 
-import android.database.Cursor;
-import android.database.SQLException;
 import android.net.Uri;
 import android.os.AsyncTask;
 
-import org.odk.collect.android.dao.InstancesDao;
+import org.odk.collect.android.database.DatabaseInstancesRepository;
 import org.odk.collect.android.forms.FormsRepository;
 import org.odk.collect.android.instances.Instance;
 import org.odk.collect.android.instances.InstancesRepository;
 import org.odk.collect.android.listeners.InstanceUploaderListener;
 import org.odk.collect.android.preferences.keys.GeneralKeys;
 import org.odk.collect.android.preferences.source.SettingsProvider;
-import org.odk.collect.android.provider.InstanceProviderAPI.InstanceColumns;
-import org.odk.collect.android.utilities.ApplicationConstants;
-import org.odk.collect.android.utilities.InstanceUploaderUtils;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.Iterator;
-import java.util.List;
 import java.util.Set;
+import java.util.stream.Stream;
 
-import timber.log.Timber;
+import static org.odk.collect.android.utilities.InstanceUploaderUtils.shouldFormBeDeleted;
 
 public abstract class InstanceUploaderTask extends AsyncTask<Long, Integer, InstanceUploaderTask.Outcome> {
 
@@ -60,67 +53,25 @@ public abstract class InstanceUploaderTask extends AsyncTask<Long, Integer, Inst
                     // Delete instances that were successfully sent and that need to be deleted
                     // either because app-level auto-delete is enabled or because the form
                     // specifies it.
-                    Set<String> keys = outcome.messagesByInstanceId.keySet();
-                    Iterator<String> it = keys.iterator();
-                    int count = keys.size();
-                    while (count > 0) {
-                        String[] selectionArgs;
-                        if (count > ApplicationConstants.SQLITE_MAX_VARIABLE_NUMBER - 1) {
-                            selectionArgs = new String[
-                                    ApplicationConstants.SQLITE_MAX_VARIABLE_NUMBER];
-                        } else {
-                            selectionArgs = new String[count + 1];
-                        }
+                    Set<String> instanceIds = outcome.messagesByInstanceId.keySet();
 
-                        StringBuilder selection = new StringBuilder();
+                    boolean isFormAutoDeleteOptionEnabled;
 
-                        selection.append(InstanceColumns._ID + " IN (");
-                        int i = 0;
-
-                        while (it.hasNext() && i < selectionArgs.length - 1) {
-                            selectionArgs[i] = it.next();
-                            selection.append('?');
-
-                            if (i != selectionArgs.length - 2) {
-                                selection.append(',');
-                            }
-                            i++;
-                        }
-
-                        count -= selectionArgs.length - 1;
-                        selection.append(") and status=?");
-                        selectionArgs[i] = Instance.STATUS_SUBMITTED;
-
-                        try (Cursor results = new InstancesDao().getInstancesCursor(null, selection.toString(), selectionArgs, null)) {
-                            if (results != null && results.getCount() > 0) {
-                                List<Long> toDelete = new ArrayList<>();
-                                results.moveToPosition(-1);
-
-                                boolean isFormAutoDeleteOptionEnabled = settingsProvider.getGeneralSettings().getBoolean(GeneralKeys.KEY_DELETE_AFTER_SEND);
-
-                                // The custom configuration from the third party app overrides
-                                // the app preferences set for delete after submission
-                                if (deleteInstanceAfterSubmission != null) {
-                                    isFormAutoDeleteOptionEnabled = deleteInstanceAfterSubmission;
-                                }
-
-                                String formId;
-                                String formVersion;
-                                while (results.moveToNext()) {
-                                    formId = results.getString(results.getColumnIndex(InstanceColumns.JR_FORM_ID));
-                                    formVersion = results.getString(results.getColumnIndex(InstanceColumns.JR_VERSION));
-                                    if (InstanceUploaderUtils.shouldFormBeDeleted(formsRepository, formId, formVersion, isFormAutoDeleteOptionEnabled)) {
-                                        toDelete.add(results.getLong(results.getColumnIndex(InstanceColumns._ID)));
-                                    }
-                                }
-
-                                DeleteInstancesTask dit = new DeleteInstancesTask(instancesRepository, formsRepository);
-                                dit.execute(toDelete.toArray(new Long[toDelete.size()]));
-                            }
-                        } catch (SQLException e) {
-                            Timber.e(e);
-                        }
+                    // The custom configuration from the third party app overrides
+                    // the app preferences set for delete after submission
+                    if (deleteInstanceAfterSubmission != null) {
+                        isFormAutoDeleteOptionEnabled = deleteInstanceAfterSubmission;
+                    } else {
+                        isFormAutoDeleteOptionEnabled = settingsProvider.getGeneralSettings().getBoolean(GeneralKeys.KEY_DELETE_AFTER_SEND);
                     }
+
+                    Stream<Instance> instancesToDelete = instanceIds.stream()
+                            .map(id -> new DatabaseInstancesRepository().get(Long.parseLong(id)))
+                            .filter(instance -> instance.getStatus().equals(Instance.STATUS_SUBMITTED))
+                            .filter(instance -> shouldFormBeDeleted(formsRepository, instance.getJrFormId(), instance.getJrVersion(), isFormAutoDeleteOptionEnabled));
+
+                    DeleteInstancesTask dit = new DeleteInstancesTask(instancesRepository, formsRepository);
+                    dit.execute(instancesToDelete.map(Instance::getId).toArray(Long[]::new));
                 }
             }
         }
@@ -155,7 +106,7 @@ public abstract class InstanceUploaderTask extends AsyncTask<Long, Integer, Inst
      * Represents the results of a submission attempt triggered by explicit user action (as opposed
      * to auto-send). A submission attempt can include finalized forms going to several different
      * servers because the app-level server configuration can be overridden by the blank form.
-     *
+     * <p>
      * The user-facing message that describes the result of a submission attempt for each specific
      * finalized form is written messages to {@link #messagesByInstanceId}. In the case of an
      * authentication request from the server, {@link #authRequestingServer} is set instead.
@@ -166,7 +117,7 @@ public abstract class InstanceUploaderTask extends AsyncTask<Long, Integer, Inst
          * attempted to be sent. This URI may not match the server specified in the app settings or
          * the blank form because there could have been a redirect. It is included in the Outcome so
          * that it can be shown to the user so s/he will know where the auth request came from.
-         *
+         * <p>
          * When this field is set, the overall submission attempt is halted so that the user can be
          * asked for credentials. Once credentials are provided, the submission attempt resumes.
          */
@@ -178,11 +129,11 @@ public abstract class InstanceUploaderTask extends AsyncTask<Long, Integer, Inst
          * submission attempt. Currently this can be either a localized message in the case of a
          * common status or an English message in the case of a rare status that is needed for
          * developer troubleshooting.
-         *
+         * <p>
          * The keys in the map are also used to identify filled forms that were part of the ongoing
          * submission attempt and don't need to be retried in the case of an authentication request.
          * See {@link #authRequestingServer}.
-         *
+         * <p>
          * TODO: Consider mapping to something machine-readable like a message ID or status ID
          * instead of a mix of localized and non-localized user-facing strings.
          */
